matrix-notifier/requirements.txt
matrix-client==0.4.0
psycopg2-binary==2.9.9
asyncio==3.4.3
